aisle prompt IS CENTERED AND SO ARE ITS CONTENTS!!!
