+++ src/components/AddAisle/AddAisle.css
@@ -1,0 +1,42 @@
+.aisle-prompt-container {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 100vw;
+}
+.aisle-prompt {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  bottom: 10%;
+  padding-top: 13px;
+  border-radius: 15px;
+  width: 80%;
+  height: 160px;
+  background-color: var(--ourPrimary);
+  max-width: 700px;
+  font-size: 1.4em;
+}
+.aisle-prompt-form {
+  margin-bottom: 10px;
+}
+.input-aisle {
+  border-radius: 15px;
+}
+#close-aisle-prompt {
+  position: absolute;
+  top: 15px;
+  right: 15px;
+}
+@media only screen and (min-width: 768px) {
+  .aisle-prompt {
+    bottom: 0;
+  }
+}
