+++ src/components/AddAisle/index.js
@@ -1,0 +1,2 @@
+import AddAisle from './AddAisle';
+export default AddAisle;
